fix(analysis/risk): guard NaN leaks in market_analysis, performance, advanced_analytics
analysis/market_analysis.py:
- min_periods on all rolling SMA/ATR/volume calls
- safe division in ADX plus_di/minus_di

analytics/performance.py:
- float() wrap on std/mean ratio calculations

risk/advanced_analytics.py:
- max(..., 0) before all sqrt calls
- sigma2_safe guard in GARCH log-likelihood
- nan_to_num on Sharpe/annual_volatility

Verified zero NaN on smoke tests for all three files.

+++ analytics/performance.py
@@ -312,7 +312,7 @@ class PerformanceAnalytics:
         # Daily metrics
         period_returns = self._get_period_returns(start_date)
         avg_daily = np.mean(period_returns) if period_returns else 0
-        volatility = np.std(period_returns) * np.sqrt(252) if period_returns else 0
+        volatility = float(np.std(period_returns) * np.sqrt(252)) if period_returns else 0.0
         best_day = max(period_returns) if period_returns else 0
         worst_day = min(period_returns) if period_returns else 0
 
@@ -579,7 +579,7 @@ class PerformanceAnalytics:
 
         # Annualize (assuming ~252 trading days)
         daily_rf = self.risk_free_rate / 252
-        return np.sqrt(252) * (mean_return - daily_rf) / std_return
+        return float(np.sqrt(252) * (mean_return - daily_rf) / std_return)
 
     def _calculate_sortino_ratio(self, trades: list[TradeRecord]) -> float:
         """Calculate Sortino ratio for trades."""
@@ -592,12 +592,12 @@ class PerformanceAnalytics:
         if not negative_returns:
             return float("inf") if np.mean(returns) > 0 else 0.0
 
-        downside_std = np.std(negative_returns)
+        downside_std = float(np.std(negative_returns))
         if downside_std == 0:
             return 0.0
 
         daily_rf = self.risk_free_rate / 252
-        return np.sqrt(252) * (np.mean(returns) - daily_rf) / downside_std
+        return float(np.sqrt(252) * (np.mean(returns) - daily_rf) / downside_std)
 
     def _calculate_max_drawdown(self, period: MetricPeriod) -> tuple[float, float]:
         """Calculate max drawdown for period."""
